Create README.md
Intranet website to manage customer food ordering and chefs’ responses on the system as the status of the order (opened, in preparing, ready, closed). Also the management for others such as cashier, stock manager, and the admin, in addition to reservation tables and some features. The aim was to replace the traditional paper system to easy, keep more comfortable for the customer and systemic system.

Technology & tools: ASP.Net MVC with SQL Server Database